% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alr}
\alias{alr}
\alias{alr.data.frame}
\alias{alr.default}
\alias{alr.matrix}
\title{alr The additive log-ratio transformation}
\usage{
alr(x, ...)

\method{alr}{default}(x, divcomp = 1, base = exp(1), removeDivComp = TRUE,
  tol = .Machine$double.eps)

\method{alr}{matrix}(x, mar = 2, divcomp = 1, base = exp(1),
  removeDivComp = TRUE, tol = .Machine$double.eps)

\method{alr}{data.frame}(x, mar = 2, ...)
}
\arguments{
\item{x}{a numeric data vector containing components of a composition}

\item{...}{additional arguments}

\item{divcomp}{index of the divisor component}

\item{base}{base of log to use, default is natural log}

\item{removeDivComp}{remove divisor component from the resulting data}

\item{tol}{machine tolerance for a zero count, default is machine tol (.Machine$double.eps)}
}
\value{
alr transformed \code{x}
}
\description{
The additive log-ratio transformation for
compositional data (not necessarily closed/normalized!)




}
\details{
The alr transformation is computed as:
 \code{x[i]} = log ( \code{x[i]} /  x[D] )
}
\examples{
# vector examples:
alr(norm_to_total(1:10))
alr(1:10)

# matrix examples:
dmat <- matrix(exp(rnorm(110)), 10)
alr(dmat, 1)
alr(dmat, 2)
}

