% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{clr}
\alias{clr}
\alias{clr.default}
\alias{clr.matrix}
\alias{clr.data.frame}
\title{clr The centered log-ratio transformation}
\usage{
clr(x, ...)

\method{clr}{default}(x, base = exp(1), tol = .Machine$double.eps)

\method{clr}{matrix}(x, mar = 2, base = exp(1), tol = .Machine$double.eps)

\method{clr}{data.frame}(x, mar = 2, ...)
}
\arguments{
\item{x}{a numeric data vector containing components of a composition}

\item{...}{additional arguments}

\item{base}{base of log to use, default is natural log}

\item{tol}{machine tolerance for a zero count, default is machine tol (.Machine$double.eps)}
}
\value{
clr transformed \code{x}
}
\description{
The centered log-ratio transformation for
compositional data (not necessarily closed/normalized!)
}
\details{
The clr is computed as
 \code{x[i]} = log (\code{x[i]} / \code{exp(mean(log(x)))})
}
\examples{
# vector examples:
clr(norm_to_total(1:10))
clr(1:10)

# matrix examples:
dmat <- matrix(exp(rnorm(110)), 10)
clr(dmat, 1)
clr(dmat, 2)
}
