% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggbiplot}
\alias{ggbiplot}
\alias{ggbiplot.default}
\alias{ggbiplot.lda}
\alias{ggbiplot.matrix}
\alias{ggbiplot.plsda}
\alias{ggbiplot.prcomp}
\alias{ggbiplot.princomp}
\alias{ggbiplot.splsda}
\title{methods for making biplots from various projection & classification models}
\usage{
ggbiplot(xobj, ...)

\method{ggbiplot}{princomp}(xobj, ...)

\method{ggbiplot}{prcomp}(xobj, ...)

\method{ggbiplot}{lda}(xobj, ...)

\method{ggbiplot}{plsda}(xobj, ...)

\method{ggbiplot}{splsda}(xobj, ...)

\method{ggbiplot}{matrix}(xobj, ...)

\method{ggbiplot}{default}(xobj, grouping, select = 1:2, circle = FALSE,
  circle.prob = 0.69, plotloadings = TRUE, scale.loadings = 1,
  col.loadings = muted("red"), alpha = 1, col = grouping,
  group.ellipse = FALSE, scale.ellipse = 1, group.cloud = FALSE,
  xlab = "", ylab = "", equalcoord = TRUE, size = 3,
  size.loadings = 1)
}
\arguments{
\item{xobj}{The object to be plotted}

\item{grouping}{an optional grouping vector (ie - for coloring points)}

\item{select}{index of components to be plotted (must be length 2)}

\item{circle}{enclose points in a circle}

\item{circle.prob}{controls circle diameter (scales data std dev) if \code{circle = TRUE}}

\item{plotloadings}{should loading vectors be plotted}

\item{scale.loadings}{scale length of loading vectors for plotting purposes}

\item{col.loadings}{a single value of vector for color of loadings}

\item{alpha}{controls relative transparency of various plot features}

\item{col}{color factor for points}

\item{group.ellipse}{enclose within-group points in an covariance ellipse}

\item{scale.ellipse}{scale \code{group.ellipse} to 1 standard deviation}

\item{group.cloud}{connect within-group points to a group mean point with a straight edge}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{equalcoord}{equal coordinates, ie should the plot area be square?}

\item{size}{point size}

\item{size.loadings}{line width of loading vectors}
}
\description{
Pretty biplots using ggplots
}
\details{
additional plotting attributes (eg colors, themes, etc) can be chained on in the usual way for ggplots
}
\examples{
# an LDA example with iris data
ldamod <- lda(iris[,1:4], grouping=iris[,5])
ggbiplot(ldamod, grouping=iris[,5]) + theme_bw(q)
}

