% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stability.R
\name{spls.stars}
\alias{spls.stars}
\title{stability selection of sparse models
' via stars}
\usage{
spls.stars(x, y, fold = 10, K, eta, kappa = 0.5, select = "pls2",
  fit = "simpls", scale.x = TRUE, scale.y = FALSE, stars.thresh = 0.05,
  ncores = 2, rep.num = 20, stars.subsample.ratio = NULL)
}
\description{
stability selection of sparse models
' via stars
}

