% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLSDA.R
\name{splsDA}
\alias{splsDA}
\title{splsDA sparse partial least squares discriminant analysis}
\usage{
splsDA(x, grouping, eta, K, usePriors = FALSE, ...)
}
\arguments{
\item{x}{x with samples in rows, features are columns (not necessarily compositional x)}

\item{grouping}{a numeric vector or factor with sample classes (length should equal \code{nrow(x)})}

\item{eta}{parameter that adjusts sparsity of the PLS model (between 0 and 1)}

\item{K}{number of components in the PLS model (default: number of classes - 1)}

\item{usePriors}{use priors for very biased sample size between groups (ie - put strong penalty on misclassifying small groups)}
}
\value{
a plsda fitted model
}
\description{
sparse Partial Least Squares Discriminant Analysis
sPLS regression to discriminate classes (via a logistic model)
basically this is a wrapper for the \code{splsda} function in the caret package,
but with default setup for dealing with uneven classes (via the priors option, see details)
see caret::splsda for implementation details
}
\details{
run this code if you don't need to fit paramaters by cross-validation
}
\seealso{
\code{\link{plsDA_main}}, \code{\link{caret::plsda}}, \code{\link{caret::splsda}}
}
