% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{plsDA}
\alias{plsDA}
\title{plsDA partial least squares discriminant analysis}
\usage{
plsDA(x, grouping, K, usePriors = FALSE, plsfun = caret::plsda, ...)
}
\arguments{
\item{x}{data with samples in rows, features are columns (not necessarily compositional data)}

\item{grouping}{a numeric vector or factor with sample classes (length should equal \code{nrow(x)})}

\item{K}{number of components in the PLS model (default: number of classes - 1)}

\item{usePriors}{use priors for very biased sample size between groups (ie - put strong penalty on misclassifying small groups)}
}
\value{
a plsda fitted model
}
\description{
Partial Least Squares Discriminant Analysis
PLS regression to discriminate classes (via a logistic model)
basically this is a wrapper for the \code{plsda} function in the caret package,
but with default setup for dealing with uneven classes (via the priors option, see details)
see caret::plsda for implementation details
}
\details{
run this code if you don't need to fit paramaters by cross-validation
}
\seealso{
\code{\link{plsDA_main}}, \code{\link{caret::plsda}}
}

